refactor(state): rename files
